Docs: Refresh integration test catalog coverage #834
